services/middleware/requirements.txt
fastapi
uvicorn[standard]
pydantic
paho-mqtt
